change timestep adder to a transformer (#1199)
