+++ pynestml/frontend/pynestml_frontend.py
@@ -61,6 +61,12 @@ def transformers_from_target_name(target_name: str, options: Optional[Mapping[st
     if options is None:
         options = {}
 
+    if target_name.upper() in ["NEST", "PYTHON_STANDALONE", "NEST_COMPARTMENTAL", "NEST_DESKTOP"]:
+        from pynestml.transformers.add_timestep_to_internals_transformer import AddTimestepToInternalsTransformer
+
+        add_timestep_to_internals_transformer = AddTimestepToInternalsTransformer()
+        transformers.append(add_timestep_to_internals_transformer)
+
     if target_name.upper() in ["NEST", "SPINNAKER"]:
         from pynestml.transformers.illegal_variable_name_transformer import IllegalVariableNameTransformer
